+++ src/main/java/com/locationbasedfoodieservice/restaurant/service/RestaurantService.java
@@ -29,7 +29,7 @@ public class RestaurantService {
     private final SigunguRepository sigunguRepository;
     private DistanceUtil distanceUtil;
     @Transactional(readOnly = true)
-    public RestaurantsResponseDto getRestaurantsBySigungu(String reqSigungu, String sort, int range) {
+    public RestaurantsResponseDto getRestaurantsBySigungu(String reqSigungu, String sort, double range) {
         //Sigungu는 "광역권,도시이름"으로 받는다고 가정 (ex : "경기,군포시")
         String[] splitSgg = reqSigungu.split(",",-1);
         String Dosi = splitSgg[0];
